Tensor has narrower type and uses imglib2 method to copy and convert.
1.
We do not depend on the RaiArrayUtils anymore.
We use the imglib2 utlity class RealTypeConverters to perform
backend copy or conversion on the fly, without any assmption
on how the backend is stored (can be an array, a list of arrays,
a cell array, or something stored on the HD).

2.
The pixel types we deal with are restricted to be RealType (no
complex values in pixel, scalar values only) and native types
(can be represented by a native type in Java: double, float, etc).

3.
Fixed the type of Tensor created or returned by copy methods.