A simple hierarchy for tensor transformation.
This hierarchy defines an interface and a utility abstract class for
tensor transformations. The tensor transformations will be
instantiated with any hyperparameters they need (e.g. the threshold
for a binarize transform).

It then defines several methods to apply this transformation to
a tensor, either in place, with a specified output tensor, or
creating a new output. Pixel type conversion on the fly is supported.